Move workload install NuGet.config file so it doesn't overwrite repo's NuGet.config (#67)
Since #66 added a NuGet.config to the root of the repo, the GH action is clobbering this file. So when we run in CI we aren't using the same NuGet.config file as a "normal" dev would use.

Moving the NuGet.config file used by `workload install` to avoid this conflict.