[net8.0] fix C# compiler error involving `x:Name` (#242)
Context: https://github.com/dotnet/maui/pull/5611
Fixes: https://github.com/microsoft/dotnet-podcasts/issues/239

The `net8.0` branch currently fails to build with:

    src/Mobile/Controls/Player.xaml.cs(54,9): error CS0103: The name 'podcastImage' does not exist in the current context
    src/Mobile/Controls/Player.xaml.cs(55,9): error CS0103: The name 'duration' does not exist in the current context

We think that `x:Name` in .NET 8 appropriately "skips" emitting fields
when used in combination with `<OnPlatform/>`.

We can condition the C# code accessing these fields with `#if`.

Secondly, there are also a lot places using `<On Platform="UWP, macOS">`, we need
to change these to:

* `macOS` -> `MacCatalyst`

Lastly, use .NET 8 RC 2 released bits. We shouldn't need to be using nightly builds in this repo.